% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_messages}
\alias{build_messages}
\title{Reconcile and Merge Messages}
\usage{
build_messages(
  summarize = TRUE,
  split = FALSE,
  as_tibble = ironmarch_as_tibble()
)
}
\arguments{
\item{summarize}{\code{logical(1L)}, Default: \code{TRUE}.
Whether to aggregate all observations into single rows, prioritizing values found in
\code{im_core_dfs}.}

\item{split}{\code{logical(1L)}, Default: \code{FALSE}.
Whether to split the result into a named list of separate data frames, retaining all
observations.
Ignored if \code{summarize} is \code{TRUE}.}

\item{as_tibble}{\code{logical(1L)}, Default: \code{TRUE}.
Whether a \code{\link[tibble:tibble]{tibble::tibble()}} should be returned. Ignored if the \code{{tibble}} package
is not installed.}
}
\value{
\itemize{
\item Data Frame
\itemize{
\item default: a \code{\link[tibble:tibble]{tibble::tibble()}} if \code{as_tibble} is \code{TRUE} and \code{{tibble}} is installed.
\item alternative: a \code{\link[data.table:data.table]{data.table::data.table()}}
}
\item If \code{split} is \code{TRUE} and \code{summarize} is \code{FALSE}, the result will be a \code{list()} of
data frames.
}
}
\description{
\code{build_messages()} reconciles and merges the messages from
\code{im_orig_dfs$orig_message_posts} and \code{im_core_dfs$core_message_posts}.
}
\examples{
# messages summarized to single rows ==================================================
build_messages()

# all message observations ============================================================
build_messages(summarize = FALSE)

# all message observations split into a list of data frames ===========================
split_messages <- build_messages(summarize = FALSE, split = TRUE)

# first five messages that appear in both... ==========================================
# ... `im_orig_dfs$orig_message_posts` and `im_core_dfs$core_message_posts`
split_messages[
  vapply(split_messages, function(.x) nrow(.x) == 2, logical(1L))
][1:5]

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
