% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo-data.R
\docType{data}
\name{combo-data}
\alias{combo-data}
\alias{combo_messages_df}
\alias{combo_members_df}
\title{Combined Tables

Data reconciled and merged from the \verb{core_*} and \verb{orig_*} tables.}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 22309 rows and 8 columns.}
\usage{
combo_messages_df

combo_members_df
}
\description{
Combined Tables

Data reconciled and merged from the \verb{core_*} and \verb{orig_*} tables.
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
\keyword{datasets}
