% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ironmarch-data.R
\docType{data}
\name{im_core_dfs}
\alias{im_core_dfs}
\alias{im_orig_dfs}
\alias{im_forums_dfs}
\alias{im_other_dfs}
\title{Iron March Data}
\format{Each object is a \code{list()} of \code{data.table::data.table()}s.}
\source{
\url{https://www.bellingcat.com/resources/how-tos/2019/11/06/massive-white-supremacist-message-board-leak-how-to-access-and-interpret-the-data/}
}
\usage{
im_core_dfs

im_orig_dfs

im_forums_dfs

im_other_dfs
}
\description{
Contents of the SQL dump from \emph{ironmarch.org}, a defunct neo-Nazi/white supremacist message board.
}
\details{
\itemize{
\item Data were acquired via the \href{https://www.bellingcat.com/}{Bellingcat}-linked \href{https://drive.google.com/file/d/1X4THLMV2Mf05hwR6Yw6IYVXhyVJt7yDg/view}{iron_march_201911_archive.torrent}.
}
}
\examples{
library(data.table)
options(datatable.print.trunc.cols = TRUE)

length(im_core_dfs)
im_core_dfs$core_members

length(im_orig_dfs)
im_orig_dfs$orig_members

length(im_forums_dfs)
im_forums_dfs$forums_posts

length(im_other_dfs)
im_other_dfs$calendar_calendars

names(im_core_dfs)

names(im_orig_dfs)

names(im_forums_dfs)

names(im_other_dfs)

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
\keyword{datasets}
