% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{im_geocode_members}
\alias{im_geocode_members}
\title{Build a Members Data Frame with Geocoded IP Addresses}
\usage{
im_geocode_members(
  all_geo_cols = TRUE,
  as_tibble = ironmarch_as_tibble(),
  as_sf = ironmarch_as_sf()
)
}
\arguments{
\item{all_geo_cols}{\code{logical(1L)}, Default \code{TRUE}.
Whether to return all georeferenced columns.}

\item{as_tibble}{\code{logical(1L)}, Default: \code{ironmarch_as_tibble()}.
Whether a \code{\link[tibble:tibble]{tibble::tibble()}} should be returned. Ignored if the \code{{tibble}} package
is not installed.}

\item{as_sf}{\code{logical(1L)}, Default: \code{ironmarch_as_sf()}.
Whether an \code{\link[sf:st_sf]{sf::st_sf()}} should be returned. Ignored if the \code{{sf}} package
is not installed.}
}
\description{
Build a Members Data Frame with Geocoded IP Addresses
}
\examples{
geocoded_members <- im_geocode_members()
geocoded_members

geocoded_cols <-  c("country", "countryCode", "region", "regionName", 
                    "city", "zip", "timezone", "isp", "org", "as")
geocoded_members[, c("name", geocoded_cols)]

plot(im_geocode_members(as_sf = FALSE)[c("lon", "lat")])

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
