% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{im_message_network}
\alias{im_message_network}
\title{Build Message Network}
\usage{
im_message_network()
}
\description{
Build Message Network
}
\examples{
im_message_g <- im_message_network()

print(im_message_g, max.lines = 20)

}
