% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ironmarch-data.R
\docType{data}
\name{iron-march-entities}
\alias{iron-march-entities}
\alias{im_groups}
\alias{im_personnel}
\alias{im_victims}
\title{Related Groups and People}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 44 rows and 8 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 1 rows and 3 columns.
}
\usage{
im_groups

im_personnel

im_victims
}
\description{
Groups and People Associated with \emph{ironmarch.org} and/or its members.
}
\examples{
im_groups
im_personnel
im_victims
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}

Alex Newhouse \email{anewhouse@miis.edu}
}
\keyword{datasets}
