% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ironmarch-data.R
\docType{data}
\name{iron-march-events}
\alias{iron-march-events}
\alias{im_events}
\title{Events related (or of likely interest), to  \emph{ironmarch.org} members.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7 rows and 4 columns.
}
\usage{
im_events
}
\description{
Events related (or of likely interest), to  \emph{ironmarch.org} members.
}
\examples{
im_events
}
\author{
Alex Newhouse \email{anewhouse@miis.edu}
}
\keyword{datasets}
