% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ironmarch-data.R
\docType{data}
\name{iron-march-third-party}
\alias{iron-march-third-party}
\alias{im_possible_mil_members}
\title{Data compiled by third parties.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 112 rows and 8 columns.
}
\source{
\itemize{
\item Bellingcat
\itemize{
\item \code{im_possible_mil_members}
\item \href{https://docs.google.com/spreadsheets/d/1mJfnYKqZdwPhK1bFbrx9hCi-61n5f0tener8iZ3AkOw/edit#gid=1584226573}{Possible Military Members (Google Sheets)}
}
}
}
\usage{
im_possible_mil_members
}
\description{
Data compiled by third parties.
}
\examples{
im_possible_mil_members
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
\keyword{datasets}
