% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{ironmarch-options}
\alias{ironmarch-options}
\alias{ironmarch_options}
\alias{ironmarch_as_tibble}
\title{\code{{ironmarch}} Options}
\usage{
ironmarch_options()

ironmarch_as_tibble(default = TRUE)
}
\arguments{
\item{default}{\code{TRUE}}
}
\description{
\code{{ironmarch}} Options
}
\examples{
ironmarch_options()

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
