% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ironmarch.R
\docType{package}
\name{ironmarch-package}
\alias{ironmarch}
\alias{ironmarch-package}
\title{ironmarch: Iron March}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Iron March.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/knapply/ironmarch}
  \item \url{https://knapply.github.io/ironmarch}
  \item Report bugs at \url{https://github.com/knapply/ironmarch/issues}
}

}
\author{
\strong{Maintainer}: Brendan Knapp \email{brendan.g.knapp@gmail.com} (\href{https://orcid.org/0000-0003-3284-4972}{ORCID})

}
\keyword{internal}
